Move all other packages and custom JS to webpacker (WIP)
