+++ Gemfile
@@ -34,9 +34,7 @@ gem "jquery-rails", ">= 4.4.0"
 gem "jquery-ui-rails", "~> 6.0", ">= 6.0.1"
 gem "jsonapi-serializer"
 gem "jwt"
-gem "local_time"
 gem "madmin"
-gem "momentjs-rails", ">= 2.20.1"
 gem "name_of_person"
 gem "normalize_country"
 gem "omniauth-facebook", ">= 8.0.0"
@@ -68,7 +66,6 @@ gem "strip_attributes"
 gem "structpluck"
 gem "terser"
 gem "turbo-rails"
-gem "vuejs-rails"
 gem "webpacker", "< 6.0"
 gem "will_paginate", "~> 3.3"
