+++ app/javascript/packs/application.js
@@ -25,10 +25,6 @@ import "chartkick/chart.js";
 import Inputmask from "inputmask/dist/jquery.inputmask";
 import "datatables.net-bs5";
 
-import TurboLinksAdapter from "vue-turbolinks";
-
-// Vue.use(TurboLinksAdapter);
-
 import { Application } from "@hotwired/stimulus"
 import { definitionsFromContext } from "@hotwired/stimulus-webpack-helpers"
