+++ app/javascript/src/event_stage/index.js
@@ -1,0 +1,1 @@
+import "./event_stage";
