+++ app/views/event_staging/events/app.html.erb
@@ -2,6 +2,8 @@
   <% "OpenSplitTime: Event staging - #{@presenter&.name || 'New event'}" %>
 <% end %>
 
+<%= javascript_pack_tag "event_stage" %>
+
 <!-- Vue.js Form Templates-->
 
 <%= render "event" %>
