+++ package.json
@@ -22,7 +22,11 @@
     "masonry-layout": "^4.2.2",
     "stimulus-use": "^0.51.1",
     "underscore": "^1.13.6",
-    "vue-turbolinks": "^2.2.2"
+    "vue": "2.5",
+    "vue-resource": "1.3",
+    "vue-router": "3.0",
+    "vue-turbolinks": "^2.2.2",
+    "vuex": "3.0"
   },
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.7.4",
